move DataValue to a new mod
